+++ Changelog.md
@@ -5,6 +5,7 @@
 - Allow ignoring updates to stop the notification on every launch
 - Update flatpak banding color
 - Track history state on more values
+- Show all history actions to undo to
 
 ## v0.9.1 &mdash; August 28th, 2026